Refactor `ApplianceEndpoints` God Class into Granular Mixins
- Created `custom_components/meraki_ha/core/api/endpoints/appliance/` package.
- Split `ApplianceEndpoints` into `ApplianceTrafficMixin` (traffic.py), `ApplianceFirewallMixin` (firewall.py), `ApplianceUplinkMixin` (uplink.py), `ApplianceVpnMixin` (vpn.py), and `ApplianceSettingsMixin` (settings.py).
- Composed `ApplianceEndpoints` in `__init__.py` inheriting from these mixins.
- Deleted the original monolithic file.
- Verified with `pytest tests/core/api/endpoints/test_appliance.py`.
